Рефакторинг: расстановка const, перенос реализации из h-файлов

+++ extensions/SharedMemory/SharedMemory.cc
@@ -587,7 +587,7 @@ namespace uniset
 			bool ok = false;
 			tm.consumer = it;
 
-			for( unsigned int i = 0; i < 2; i++ )
+			for( size_t i = 0; i < 2; i++ )
 			{
 				try
 				{
